enabled add project todo button
