+++ src/template.html
@@ -84,8 +84,8 @@
                       <div class="form-group">
                         <label for="todo-priority">priority</label>
                         <select class="form-control" id="todo-priority">
-                          <option>Urgent</option>
-                          <option>Normal</option>
+                          <option value="Urgent">Urgent</option>
+                          <option value="Normal">Normal</option>
                         </select>
                       </div>
                     </form>
@@ -150,19 +150,19 @@
                     <form>
                       <div class="form-group">
                         <label for="todo-title">Title</label>
-                        <input type="text" class="form-control" id="todo-title" placeholder="Enter todo name">
+                        <input type="text" class="form-control" id="edit-todo-title" placeholder="Enter todo name">
                       </div>
                       <div class="form-group">
                         <label for="todo-description">Description</label>
-                        <textarea class="form-control" id="todo-description" rows="2" placeholder="Enter Description" ></textarea>
+                        <textarea class="form-control" id="edit-todo-description" rows="2" placeholder="Enter Description" ></textarea>
                       </div>
                       <div class="form-group">
                         <label for="todo-date">Date</label>
-                        <input type="date" id="todo-date">
+                        <input type="date" id="edit-todo-date">
                       </div>
                       <div class="form-group">
                         <label for="todo-priority">priority</label>
-                        <select class="form-control" id="todo-priority">
+                        <select class="form-control" id="edit-todo-priority">
                           <option value="Urgent">Urgent</option>
                           <option value="Normal">Normal</option>
                         </select>
